Updated to stack LTS 12.11
This updates http-clint to 0.5.13.1 (from 0.5.7), which seems to resolve
a sporadic issue that status code exceptions would not be wrapped using
the provided exception wrapper function, which could result in an
exception when trying to access uncached files.

Also, Data.Monoid is now exported by the Prelude, so the explicit
imports have been removed.